whatsapp-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
python-dotenv==0.19.0
urllib3==1.26.6
aiohttp==3.9.1